+++ rerun_py/src/dataframe/mod.rs
@@ -5,6 +5,8 @@
 
 mod component_columns;
 mod index_columns;
+mod recording;
+mod recording_view;
 mod schema;
 
 use std::{
@@ -14,7 +16,7 @@ use std::{
 };
 
 use arrow::{
-    array::{ArrayData, Int64Array, RecordBatchIterator, RecordBatchReader, make_array},
+    array::{ArrayData, Int64Array, make_array},
     pyarrow::PyArrowType,
 };
 use datafusion::catalog::TableProvider;
@@ -22,27 +24,22 @@ use datafusion_ffi::table_provider::FFI_TableProvider;
 use numpy::PyArrayMethods as _;
 use pyo3::{
     exceptions::{PyRuntimeError, PyTypeError, PyValueError},
-    prelude::*,
-    types::{PyCapsule, PyDict, PyTuple},
+    prelude::*, //TODO remove this
+    types::PyCapsule,
 };
 
 use re_arrow_util::ArrowArrayDowncastRef as _;
-use re_chunk::ComponentIdentifier;
-use re_chunk_store::{
-    ChunkStore, ChunkStoreConfig, ChunkStoreHandle, ColumnDescriptor, QueryExpression,
-    SparseFillStrategy, StaticColumnSelection, ViewContentsSelector,
-};
-use re_dataframe::{QueryEngine, StorageEngine};
-use re_log_types::{EntityPathFilter, ResolvedTimeRange};
+use re_chunk_store::{ChunkStore, ChunkStoreConfig, ChunkStoreHandle, ColumnDescriptor};
 use re_sdk::{StoreId, StoreKind};
 use re_sorbet::{ColumnSelector, ComponentColumnSelector, TimeColumnSelector};
 
 pub use self::{
     component_columns::{PyComponentColumnDescriptor, PyComponentColumnSelector},
     index_columns::{PyIndexColumnDescriptor, PyIndexColumnSelector},
+    recording::{PyRecording, PyRecordingHandle},
+    recording_view::PyRecordingView,
     schema::PySchema,
 };
-use super::utils::py_rerun_warn_cstr;
 use crate::{catalog::PyCatalogClientInternal, utils::get_tokio_runtime};
 
 /// Register the `rerun.dataframe` module.
@@ -231,817 +228,6 @@ impl IndexValuesLike<'_> {
                 }
             }
         }
-    }
-}
-
-/// A single Rerun recording.
-///
-/// This can be loaded from an RRD file using [`load_recording()`][rerun.dataframe.load_recording].
-///
-/// A recording is a collection of data that was logged to Rerun. This data is organized
-/// as a column for each index (timeline) and each entity/component pair that was logged.
-///
-/// You can examine the [`.schema()`][rerun.dataframe.Recording.schema] of the recording to see
-/// what data is available, or create a [`RecordingView`][rerun.dataframe.RecordingView] to
-/// to retrieve the data.
-#[pyclass(name = "Recording")]
-pub struct PyRecording {
-    pub(crate) store: ChunkStoreHandle,
-    pub(crate) cache: re_dataframe::QueryCacheHandle,
-}
-
-#[derive(Clone)]
-pub enum PyRecordingHandle {
-    Local(std::sync::Arc<Py<PyRecording>>),
-    // TODO(rerun-io/dataplatform#405): interface with remote data needs to be reimplemented
-    //Remote(std::sync::Arc<Py<PyRemoteRecording>>),
-}
-
-/// A view of a recording restricted to a given index, containing a specific set of entities and components.
-///
-/// See [`Recording.view(…)`][rerun.dataframe.Recording.view] for details on how to create a `RecordingView`.
-///
-/// Note: `RecordingView` APIs never mutate the underlying view. Instead, they
-/// always return new views with the requested modifications applied.
-///
-/// The view will only contain a single row for each unique value of the index
-/// that is associated with a component column that was included in the view.
-/// Component columns that are not included via the view contents will not
-/// impact the rows that make up the view. If the same entity / component pair
-/// was logged to a given index multiple times, only the most recent row will be
-/// included in the view, as determined by the `row_id` column. This will
-/// generally be the last value logged, as row_ids are guaranteed to be
-/// monotonically increasing when data is sent from a single process.
-#[pyclass(name = "RecordingView")]
-#[derive(Clone)]
-pub struct PyRecordingView {
-    pub(crate) recording: PyRecordingHandle,
-
-    pub(crate) query_expression: QueryExpression,
-}
-
-impl PyRecordingView {
-    fn select_args(
-        args: &Bound<'_, PyTuple>,
-        columns: Option<Vec<AnyColumn>>,
-    ) -> PyResult<Option<Vec<ColumnSelector>>> {
-        // Coerce the arguments into a list of `ColumnSelector`s
-        let args: Vec<AnyColumn> = args
-            .iter()
-            .map(|arg| arg.extract::<AnyColumn>())
-            .collect::<PyResult<_>>()?;
-
-        if columns.is_some() && !args.is_empty() {
-            return Err(PyValueError::new_err(
-                "Cannot specify both `columns` and `args` in `select`.",
-            ));
-        }
-
-        let columns = columns.or(if !args.is_empty() { Some(args) } else { None });
-
-        columns
-            .map(|cols| {
-                cols.into_iter()
-                    .map(|col| col.into_selector())
-                    .collect::<PyResult<_>>()
-            })
-            .transpose()
-    }
-}
-
-/// A view of a recording restricted to a given index, containing a specific set of entities and components.
-///
-/// Can only be created by calling `view(...)` on a `Recording`.
-///
-/// The only type of index currently supported is the name of a timeline.
-///
-/// The view will only contain a single row for each unique value of the index. If the same entity / component pair
-/// was logged to a given index multiple times, only the most recent row will be included in the view, as determined
-/// by the `row_id` column. This will generally be the last value logged, as row_ids are guaranteed to be monotonically
-/// increasing when data is sent from a single process.
-#[pymethods]
-impl PyRecordingView {
-    /// The schema describing all the columns available in the view.
-    ///
-    /// This schema will only contain the columns that are included in the view via
-    /// the view contents.
-    fn schema(&self, py: Python<'_>) -> PySchema {
-        match &self.recording {
-            PyRecordingHandle::Local(recording) => {
-                let borrowed: PyRef<'_, PyRecording> = recording.borrow(py);
-                let engine = borrowed.engine();
-
-                let mut query_expression = self.query_expression.clone();
-                query_expression.selection = None;
-
-                PySchema {
-                    schema: engine.schema_for_query(&query_expression).into(),
-                }
-            }
-        }
-    }
-
-    /// Select the columns from the view.
-    ///
-    /// If no columns are provided, all available columns will be included in
-    /// the output.
-    ///
-    /// The selected columns do not change the rows that are included in the
-    /// view. The rows are determined by the index values and the components
-    /// that were included in the view contents, or can be overridden with
-    /// [`.using_index_values()`][rerun.dataframe.RecordingView.using_index_values].
-    ///
-    /// If a column was not provided with data for a given row, it will be
-    /// `null` in the output.
-    ///
-    /// The output is a [`pyarrow.RecordBatchReader`][] that can be used to read
-    /// out the data.
-    ///
-    /// Parameters
-    /// ----------
-    /// *args : AnyColumn
-    ///     The columns to select.
-    /// columns : Optional[Sequence[AnyColumn]], optional
-    ///     Alternatively the columns to select can be provided as a sequence.
-    ///
-    /// Returns
-    /// -------
-    /// pa.RecordBatchReader
-    ///     A reader that can be used to read out the selected data.
-    #[pyo3(signature = (
-        *args,
-        columns = None
-    ))]
-    fn select(
-        &self,
-        py: Python<'_>,
-        args: &Bound<'_, PyTuple>,
-        columns: Option<Vec<AnyColumn>>,
-    ) -> PyResult<PyArrowType<Box<dyn RecordBatchReader + Send>>> {
-        let mut query_expression = self.query_expression.clone();
-        query_expression.selection = Self::select_args(args, columns)?;
-
-        match &self.recording {
-            PyRecordingHandle::Local(recording) => {
-                let borrowed = recording.borrow(py);
-                let engine = borrowed.engine();
-
-                let query_handle = engine.query(query_expression);
-
-                // If the only contents found are static, we might need to warn the user since
-                // this means we won't naturally have any rows in the result.
-                let available_data_columns = &query_handle.view_contents().components;
-
-                // We only consider all contents static if there at least some columns
-                let all_contents_are_static = !available_data_columns.is_empty()
-                    && available_data_columns.iter().all(|c| c.is_static());
-
-                // Additionally, we only want to warn if the user actually tried to select some
-                // of the static columns. Otherwise the fact that there are no results shouldn't
-                // be surprising.
-                let selected_data_columns = query_handle
-                    .selected_contents()
-                    .iter()
-                    .map(|(_, col)| col)
-                    .filter(|c| matches!(c, ColumnDescriptor::Component(_)))
-                    .collect::<Vec<_>>();
-
-                let any_selected_data_is_static =
-                    selected_data_columns.iter().any(|c| c.is_static());
-
-                if self.query_expression.using_index_values.is_none()
-                    && all_contents_are_static
-                    && any_selected_data_is_static
-                    && self.query_expression.filtered_index.is_some()
-                {
-                    py_rerun_warn_cstr(c"RecordingView::select: tried to select static data, but no non-static contents generated an index value on this timeline. No results will be returned. Either include non-static data or consider using `select_static()` instead.")?;
-                }
-
-                let schema = query_handle.schema().clone();
-
-                let reader =
-                    RecordBatchIterator::new(query_handle.into_batch_iter().map(Ok), schema);
-                Ok(PyArrowType(Box::new(reader)))
-            }
-        }
-    }
-
-    /// Select only the static columns from the view.
-    ///
-    /// Because static data has no associated index values it does not cause a
-    /// row to be generated in the output. If your view only contains static data
-    /// this method allows you to select it without needing to provide index values.
-    ///
-    /// This method will always return a single row.
-    ///
-    /// Any non-static columns that are included in the selection will generate a warning
-    /// and produce empty columns.
-    ///
-    ///
-    /// Parameters
-    /// ----------
-    /// *args : AnyColumn
-    ///     The columns to select.
-    /// columns : Optional[Sequence[AnyColumn]], optional
-    ///     Alternatively the columns to select can be provided as a sequence.
-    ///
-    /// Returns
-    /// -------
-    /// pa.RecordBatchReader
-    ///     A reader that can be used to read out the selected data.
-    //TODO(#10335): remove deprecated method
-    #[pyo3(signature = (
-        *args,
-        columns = None
-    ))]
-    fn select_static(
-        &self,
-        py: Python<'_>,
-        args: &Bound<'_, PyTuple>,
-        columns: Option<Vec<AnyColumn>>,
-    ) -> PyResult<PyArrowType<Box<dyn RecordBatchReader + Send>>> {
-        let mut query_expression = self.query_expression.clone();
-        // This is a static selection, so we clear the filtered index
-        query_expression.filtered_index = None;
-
-        //TODO(#10327): this should not be necessary!
-        query_expression.sparse_fill_strategy = SparseFillStrategy::LatestAtGlobal;
-
-        // If no columns provided, select all static columns
-        let static_columns = Self::select_args(args, columns)
-            .transpose()
-            .unwrap_or_else(|| {
-                Ok(self
-                    .schema(py)
-                    .schema
-                    .component_columns()
-                    .filter(|col| col.is_static())
-                    .map(|col| ColumnDescriptor::Component(col.clone()).into())
-                    .collect())
-            })?;
-
-        query_expression.selection = Some(static_columns);
-
-        match &self.recording {
-            PyRecordingHandle::Local(recording) => {
-                let borrowed = recording.borrow(py);
-                let engine = borrowed.engine();
-
-                let query_handle = engine.query(query_expression);
-
-                let non_static_cols = query_handle
-                    .selected_contents()
-                    .iter()
-                    .filter(|(_, col)| !col.is_static())
-                    .collect::<Vec<_>>();
-
-                if !non_static_cols.is_empty() {
-                    return Err(PyValueError::new_err(format!(
-                        "Static selection resulted in non-static columns: {non_static_cols:?}",
-                    )));
-                }
-
-                let schema = query_handle.schema().clone();
-
-                let reader =
-                    RecordBatchIterator::new(query_handle.into_batch_iter().map(Ok), schema);
-
-                Ok(PyArrowType(Box::new(reader)))
-            }
-        }
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Filter the view to only include data between the given index sequence numbers.
-    ///
-    /// This range is inclusive and will contain both the value at the start and the value at the end.
-    ///
-    /// The view must be of a sequential index type to use this method.
-    ///
-    /// Parameters
-    /// ----------
-    /// start : int
-    ///     The inclusive start of the range.
-    /// end : int
-    ///     The inclusive end of the range.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing only the data within the specified range.
-    ///
-    ///     The original view will not be modified.
-    fn filter_range_sequence(&self, start: i64, end: i64) -> PyResult<Self> {
-        match self.query_expression.filtered_index.as_ref() {
-            // TODO(#9084): do we need this check? If so, how can we accomplish it?
-            // Some(filtered_index) if filtered_index.typ() != TimeType::Sequence => {
-            //     return Err(PyValueError::new_err(format!(
-            //         "Index for {} is not a sequence.",
-            //         filtered_index.name()
-            //     )));
-            // }
-            Some(_) => {}
-
-            None => {
-                return Err(PyValueError::new_err(
-                    "Specify an index to filter on first.".to_owned(),
-                ));
-            }
-        }
-
-        let start = if let Ok(seq) = re_chunk::TimeInt::try_from(start) {
-            seq
-        } else {
-            re_log::error!(
-                illegal_value = start,
-                new_value = re_chunk::TimeInt::MIN.as_i64(),
-                "set_time_sequence() called with illegal value - clamped to minimum legal value"
-            );
-            re_chunk::TimeInt::MIN
-        };
-
-        let end = if let Ok(seq) = re_chunk::TimeInt::try_from(end) {
-            seq
-        } else {
-            re_log::error!(
-                illegal_value = end,
-                new_value = re_chunk::TimeInt::MAX.as_i64(),
-                "set_time_sequence() called with illegal value - clamped to maximum legal value"
-            );
-            re_chunk::TimeInt::MAX
-        };
-
-        let resolved = ResolvedTimeRange::new(start, end);
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.filtered_index_range = Some(resolved);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Filter the view to only include data between the given index values expressed as seconds.
-    ///
-    /// This range is inclusive and will contain both the value at the start and the value at the end.
-    ///
-    /// The view must be of a temporal index type to use this method.
-    ///
-    /// Parameters
-    /// ----------
-    /// start : int
-    ///     The inclusive start of the range.
-    /// end : int
-    ///     The inclusive end of the range.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing only the data within the specified range.
-    ///
-    ///     The original view will not be modified.
-    fn filter_range_secs(&self, start: f64, end: f64) -> PyResult<Self> {
-        match self.query_expression.filtered_index.as_ref() {
-            // TODO(#9084): do we need this check? If so, how can we accomplish it?
-            // Some(filtered_index) if filtered_index.typ() != TimeType::Time => {
-            //     return Err(PyValueError::new_err(format!(
-            //         "Index for {} is not temporal.",
-            //         filtered_index.name()
-            //     )));
-            // }
-            Some(_) => {}
-
-            None => {
-                return Err(PyValueError::new_err(
-                    "Specify an index to filter on first.".to_owned(),
-                ));
-            }
-        }
-
-        let start = re_log_types::Timestamp::from_secs_since_epoch(start);
-        let end = re_log_types::Timestamp::from_secs_since_epoch(end);
-
-        let resolved = ResolvedTimeRange::new(start, end);
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.filtered_index_range = Some(resolved);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    /// DEPRECATED: Renamed to `filter_range_secs`.
-    #[deprecated(since = "0.23.0", note = "Renamed to `filter_range_secs`")]
-    fn filter_range_seconds(&self, start: f64, end: f64) -> PyResult<Self> {
-        self.filter_range_secs(start, end)
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Filter the view to only include data between the given index values expressed as nanoseconds.
-    ///
-    /// This range is inclusive and will contain both the value at the start and the value at the end.
-    ///
-    /// The view must be of a temporal index type to use this method.
-    ///
-    /// Parameters
-    /// ----------
-    /// start : int
-    ///     The inclusive start of the range.
-    /// end : int
-    ///     The inclusive end of the range.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing only the data within the specified range.
-    ///
-    ///     The original view will not be modified.
-    fn filter_range_nanos(&self, start: i64, end: i64) -> PyResult<Self> {
-        match self.query_expression.filtered_index.as_ref() {
-            // TODO(#9084): do we need this?
-            // Some(filtered_index) if filtered_index.typ() != TimeType::Time => {
-            //     return Err(PyValueError::new_err(format!(
-            //         "Index for {} is not temporal.",
-            //         filtered_index.name()
-            //     )));
-            // }
-            Some(_) => {}
-
-            None => {
-                return Err(PyValueError::new_err(
-                    "Specify an index to filter on first.".to_owned(),
-                ));
-            }
-        }
-
-        let start = re_log_types::Timestamp::from_nanos_since_epoch(start);
-        let end = re_log_types::Timestamp::from_nanos_since_epoch(end);
-
-        let resolved = ResolvedTimeRange::new(start, end);
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.filtered_index_range = Some(resolved);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Filter the view to only include data at the provided index values.
-    ///
-    /// The index values returned will be the intersection between the provided values and the
-    /// original index values.
-    ///
-    /// This requires index values to be a precise match. Index values in Rerun are
-    /// represented as i64 sequence counts or nanoseconds. This API does not expose an interface
-    /// in floating point seconds, as the numerical conversion would risk false mismatches.
-    ///
-    /// Parameters
-    /// ----------
-    /// values : IndexValuesLike
-    ///     The index values to filter by.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing only the data at the specified index values.
-    ///
-    ///     The original view will not be modified.
-    fn filter_index_values(&self, values: IndexValuesLike<'_>) -> PyResult<Self> {
-        let values = values.to_index_values()?;
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.filtered_index_values = Some(values);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Filter the view to only include rows where the given component column is not null.
-    ///
-    /// This corresponds to rows for index values where this component was provided to Rerun explicitly
-    /// via `.log()` or `.send_columns()`.
-    ///
-    /// Parameters
-    /// ----------
-    /// column : AnyComponentColumn
-    ///     The component column to filter by.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing only the data where the specified component column is not null.
-    ///
-    ///     The original view will not be modified.
-    fn filter_is_not_null(&self, column: AnyComponentColumn) -> PyResult<Self> {
-        let column = column.into_selector();
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.filtered_is_not_null = Some(column?);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Replace the index in the view with the provided values.
-    ///
-    /// The output view will always have the same number of rows as the provided values, even if
-    /// those rows are empty. Use with [`.fill_latest_at()`][rerun.dataframe.RecordingView.fill_latest_at]
-    /// to populate these rows with the most recent data.
-    ///
-    /// This requires index values to be a precise match. Index values in Rerun are
-    /// represented as i64 sequence counts or nanoseconds. This API does not expose an interface
-    /// in floating point seconds, as the numerical conversion would risk false mismatches.
-    ///
-    /// Parameters
-    /// ----------
-    /// values : IndexValuesLike
-    ///     The index values to use.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view containing the provided index values.
-    ///
-    ///     The original view will not be modified.
-    fn using_index_values(&self, values: IndexValuesLike<'_>) -> PyResult<Self> {
-        let values = values.to_index_values()?;
-
-        let mut query_expression = self.query_expression.clone();
-        query_expression.using_index_values = Some(values);
-
-        Ok(Self {
-            recording: self.recording.clone(),
-            query_expression,
-        })
-    }
-
-    #[allow(rustdoc::private_doc_tests)]
-    /// Populate any null values in a row with the latest valid data according to the index.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     A new view with the null values filled in.
-    ///
-    ///     The original view will not be modified.
-    fn fill_latest_at(&self) -> Self {
-        let mut query_expression = self.query_expression.clone();
-        query_expression.sparse_fill_strategy = SparseFillStrategy::LatestAtGlobal;
-
-        Self {
-            recording: self.recording.clone(),
-            query_expression,
-        }
-    }
-}
-
-impl PyRecording {
-    fn engine(&self) -> QueryEngine<StorageEngine> {
-        // Safety: this is all happening in the context of a python client using the dataframe API,
-        // there is no reason to worry about handle leakage whatsoever.
-        #[allow(unsafe_code)]
-        let engine = unsafe { StorageEngine::new(self.store.clone(), self.cache.clone()) };
-
-        QueryEngine { engine }
-    }
-
-    /// Convert a `ViewContentsLike` into a `ViewContentsSelector`.
-    ///
-    /// ```python
-    /// ViewContentsLike = Union[str, Dict[str, Union[str, Sequence[str]]]]
-    /// ```
-    ///
-    /// We cant do this with the normal `FromPyObject` mechanisms because we want access to the
-    /// `QueryEngine` to resolve the entity paths.
-    fn extract_contents_expr(
-        &self,
-        expr: Bound<'_, PyAny>,
-    ) -> PyResult<re_chunk_store::ViewContentsSelector> {
-        let engine = self.engine();
-
-        if let Ok(expr) = expr.extract::<String>() {
-            // `str`
-
-            let path_filter =
-                EntityPathFilter::parse_strict(&expr)
-                    .map_err(|err| {
-                        PyValueError::new_err(format!(
-                            "Could not interpret `contents` as a ViewContentsLike. Failed to parse {expr}: {err}.",
-                        ))
-                    })?;
-
-            let contents = engine
-                .iter_entity_paths_sorted(&path_filter)
-                .map(|p| (p, None))
-                .collect();
-
-            Ok(contents)
-        } else if let Ok(dict) = expr.downcast::<PyDict>() {
-            // `Union[str, Sequence[str]]]`
-
-            let mut contents = ViewContentsSelector::default();
-
-            for (key, value) in dict {
-                let key = key.extract::<String>().map_err(|_err| {
-                    PyTypeError::new_err(
-                        format!("Could not interpret `contents` as a ViewContentsLike. Key: {key} is not a path expression."),
-                    )
-                })?;
-
-                let path_filter = EntityPathFilter::parse_strict(&key).map_err(|err| {
-                    PyValueError::new_err(format!(
-                        "Could not interpret `contents` as a ViewContentsLike. Failed to parse {key}: {err}.",
-                    ))
-                })?;
-
-                let component_strs: BTreeSet<ComponentIdentifier> = if let Ok(component) =
-                    value.extract::<String>()
-                {
-                    std::iter::once(component.into()).collect()
-                } else if let Ok(components) = value.extract::<Vec<String>>() {
-                    components.into_iter().map(Into::into).collect()
-                } else {
-                    return Err(PyTypeError::new_err(format!(
-                        "Could not interpret `contents` as a ViewContentsLike. Value: {value} is not a str or Sequence[str]."
-                    )));
-                };
-
-                contents.append(
-                    &mut engine
-                        .iter_entity_paths_sorted(&path_filter)
-                        .map(|entity_path| (entity_path, Some(component_strs.clone())))
-                        .collect(),
-                );
-            }
-
-            Ok(contents)
-        } else {
-            return Err(PyTypeError::new_err(
-                "Could not interpret `contents` as a ViewContentsLike. Top-level type must be a string or a dictionary.",
-            ));
-        }
-    }
-}
-
-#[pymethods]
-impl PyRecording {
-    /// The schema describing all the columns available in the recording.
-    fn schema(&self) -> PySchema {
-        PySchema {
-            schema: self.store.read().schema().into(),
-        }
-    }
-
-    #[allow(rustdoc::private_doc_tests, rustdoc::invalid_rust_codeblocks)]
-    /// Create a [`RecordingView`][rerun.dataframe.RecordingView] of the recording according to a particular index and content specification.
-    ///
-    /// The only type of index currently supported is the name of a timeline, or `None` (see below
-    /// for details).
-    ///
-    /// The view will only contain a single row for each unique value of the index
-    /// that is associated with a component column that was included in the view.
-    /// Component columns that are not included via the view contents will not
-    /// impact the rows that make up the view. If the same entity / component pair
-    /// was logged to a given index multiple times, only the most recent row will be
-    /// included in the view, as determined by the `row_id` column. This will
-    /// generally be the last value logged, as row_ids are guaranteed to be
-    /// monotonically increasing when data is sent from a single process.
-    ///
-    /// If `None` is passed as the index, the view will contain only static columns (among those
-    /// specified) and no index columns. It will also contain a single row per partition.
-    ///
-    /// Parameters
-    /// ----------
-    /// index : str | None
-    ///     The index to use for the view. This is typically a timeline name. Use `None` to query static data only.
-    /// contents : ViewContentsLike
-    ///     The content specification for the view.
-    ///
-    ///     This can be a single string content-expression such as: `"world/cameras/**"`, or a dictionary
-    ///     specifying multiple content-expressions and a respective list of components to select within
-    ///     that expression such as `{"world/cameras/**": ["ImageBuffer", "PinholeProjection"]}`.
-    /// include_semantically_empty_columns : bool, optional
-    ///     Whether to include columns that are semantically empty, by default `False`.
-    ///
-    ///     Semantically empty columns are components that are `null` or empty `[]` for every row in the recording.
-    /// include_indicator_columns : bool, optional
-    ///     Whether to include indicator columns, by default `False`.
-    ///
-    ///     Indicator columns are components used to represent the presence of an archetype within an entity.
-    /// include_tombstone_columns : bool, optional
-    ///     Whether to include tombstone columns, by default `False`.
-    ///
-    ///     Tombstone columns are components used to represent clears. However, even without the clear
-    ///     tombstone columns, the view will still apply the clear semantics when resolving row contents.
-    ///
-    /// Returns
-    /// -------
-    /// RecordingView
-    ///     The view of the recording.
-    ///
-    /// Examples
-    /// --------
-    /// All the data in the recording on the timeline "my_index":
-    /// ```python
-    /// recording.view(index="my_index", contents="/**")
-    /// ```
-    ///
-    /// Just the Position3D components in the "points" entity:
-    /// ```python
-    /// recording.view(index="my_index", contents={"points": "Position3D"})
-    /// ```
-    #[allow(clippy::fn_params_excessive_bools)]
-    #[pyo3(signature = (
-        *,
-        index,
-        contents,
-        include_semantically_empty_columns = false,
-        include_indicator_columns = false,
-        include_tombstone_columns = false,
-    ))]
-    fn view(
-        slf: Bound<'_, Self>,
-        index: Option<&str>,
-        contents: Bound<'_, PyAny>,
-        include_semantically_empty_columns: bool,
-        include_indicator_columns: bool,
-        include_tombstone_columns: bool,
-    ) -> PyResult<PyRecordingView> {
-        let static_only = index.is_none();
-
-        let borrowed_self = slf.borrow();
-
-        // Look up the type of the timeline
-        let filtered_index = index.map(|index| {
-            let selector = TimeColumnSelector::from(index);
-            let time_column = borrowed_self.store.read().resolve_time_selector(&selector);
-            *time_column.timeline().name()
-        });
-
-        let contents = borrowed_self.extract_contents_expr(contents)?;
-
-        let query = QueryExpression {
-            view_contents: Some(contents),
-            include_semantically_empty_columns,
-            include_indicator_columns,
-            include_tombstone_columns,
-            include_static_columns: if static_only {
-                StaticColumnSelection::StaticOnly
-            } else {
-                StaticColumnSelection::Both
-            },
-            filtered_index,
-            filtered_index_range: None,
-            filtered_index_values: None,
-            using_index_values: None,
-            filtered_is_not_null: None,
-            //TODO(#10327): this should not be necessary!
-            sparse_fill_strategy: if static_only {
-                SparseFillStrategy::LatestAtGlobal
-            } else {
-                SparseFillStrategy::None
-            },
-            selection: None,
-        };
-
-        let recording = slf.unbind();
-
-        Ok(PyRecordingView {
-            recording: PyRecordingHandle::Local(std::sync::Arc::new(recording)),
-            query_expression: query,
-        })
-    }
-
-    /// The recording ID of the recording.
-    fn recording_id(&self) -> String {
-        self.store.read().id().as_str().to_owned()
-    }
-
-    /// The application ID of the recording.
-    fn application_id(&self) -> PyResult<String> {
-        Ok(self
-            .store
-            .read()
-            .store_info()
-            .ok_or(PyValueError::new_err(
-                "Recording is missing application id.",
-            ))?
-            .application_id
-            .as_str()
-            .to_owned())
     }
 }
